promote plugin api v3 author surface
