+++ Cargo.toml
@@ -31,8 +31,8 @@ lix_slatedb_storage = { path = "packages/slatedb-storage", version = "0.9.0" }
 lix_sqlite_storage = { path = "packages/sqlite-storage", version = "0.9.0" }
 lix_js_sdk = { path = "packages/js-sdk", version = "0.9.0" }
 lix_order_key = { path = "plugins/order-key", version = "0.9.0" }
-lix_plugin_api_v2 = { path = "packages/plugin-api", version = "0.9.0" }
-lix_plugin_api_v3_prototype = { path = "packages/plugin-api-v3-prototype", version = "0.9.0" }
+lix_plugin_api_v2 = { path = "packages/plugin-api-v2", version = "0.9.0" }
+lix_plugin_api = { path = "packages/plugin-api", version = "0.9.0" }
 lix_plugin_arena = { path = "packages/plugin-arena", version = "0.9.0" }
 lix_sdk = { path = "packages/rs-sdk", version = "0.9.0" }
 lix_server_protocol = { path = "packages/server-protocol", version = "0.9.0" }
