+++ packages/plugin-api-v2/Cargo.toml
@@ -1,27 +1,22 @@
 [package]
-name = "plugin_csv_v3_prototype"
+name = "lix_plugin_api_v2"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
-description = "CSV prototype for the fused Lix Wasm Component API v3"
+description = "Author-facing Rust API for Lix Wasm Component v2 plugins"
 license.workspace = true
 homepage.workspace = true
 repository.workspace = true
-documentation.workspace = true
+documentation = "https://docs.rs/lix_plugin_api_v2"
 readme = "README.md"
 keywords.workspace = true
 categories.workspace = true
 rust-version.workspace = true
-publish = false
 
 [lints]
 workspace = true
 
-[lib]
-crate-type = ["cdylib", "rlib"]
-
 [dependencies]
-base64 = "0.22"
-lix_plugin_api_v3_prototype = { workspace = true }
 serde_json = "1"
 uuid = { version = "1", features = ["std"] }
+wit-bindgen = "0.57"
