+++ packages/rs-sdk-tests/Cargo.toml
@@ -16,13 +16,13 @@ lix_sdk = { path = "../rs-sdk", version = "0.9.0", default-features = false, fea
 lix_engine = { path = "../engine", version = "0.9.0", features = ["storage-benches"] }
 lix_rocksdb_storage = { path = "../rocksdb-storage", version = "0.9.0" }
 plugin_csv_v2 = { path = "../../plugins/csv-v2", lib = true, artifact = "cdylib", target = "wasm32-wasip2" }
-plugin_csv_v3_prototype = { path = "../../plugins/csv-v3-prototype", lib = true, artifact = "cdylib", target = "wasm32-wasip2" }
-plugin_excalidraw_v3_prototype = { path = "../../plugins/excalidraw-v3-prototype", lib = true, artifact = "cdylib", target = "wasm32-wasip2" }
+plugin_csv = { path = "../../plugins/csv", lib = true, artifact = "cdylib", target = "wasm32-wasip2" }
+plugin_excalidraw = { path = "../../plugins/excalidraw", lib = true, artifact = "cdylib", target = "wasm32-wasip2" }
 plugin_markdown_incremental_v2 = { path = "../../plugins/markdown-v2", lib = true, artifact = "cdylib", target = "wasm32-wasip2" }
-plugin_markdown_v3_prototype = { path = "../../plugins/markdown-v3-prototype", lib = true, artifact = "cdylib", target = "wasm32-wasip2" }
+plugin_markdown = { path = "../../plugins/markdown", lib = true, artifact = "cdylib", target = "wasm32-wasip2" }
 plugin_excalidraw_v2 = { path = "../../plugins/excalidraw-v2", lib = true, artifact = "cdylib", target = "wasm32-wasip2" }
 plugin_json_incremental_v2 = { path = "../../plugins/json-v2", lib = true, artifact = "cdylib", target = "wasm32-wasip2" }
-plugin_json_v3_prototype = { path = "../../plugins/json-v3-prototype", lib = true, artifact = "cdylib", target = "wasm32-wasip2" }
+plugin_json = { path = "../../plugins/json", lib = true, artifact = "cdylib", target = "wasm32-wasip2" }
 plugin_git_text_v2 = { path = "../../plugins/text-v2", lib = true, artifact = "cdylib", target = "wasm32-wasip2" }
 
 async-trait = "0.1"
