+++ plugins/csv/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "lix_plugin_api_v3_prototype"
+name = "plugin_csv"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
-description = "Prototype fused push API for Lix Wasm Component plugins"
+description = "CSV plugin for the fused Lix Wasm Component API v3"
 license.workspace = true
 homepage.workspace = true
 repository.workspace = true
@@ -17,6 +17,11 @@ publish = false
 [lints]
 workspace = true
 
+[lib]
+crate-type = ["cdylib", "rlib"]
+
 [dependencies]
+base64 = "0.22"
+lix_plugin_api = { workspace = true }
+serde_json = "1"
 uuid = { version = "1", features = ["std"] }
-wit-bindgen = "0.57"
