+++ plugins/csv/README.md
@@ -1,4 +1,4 @@
-# CSV Component API v3 prototype
+# CSV Component API v3
 
 Fused, push-based initial CSV import used only for runtime and boundary
 benchmarking.
