+++ plugins/excalidraw/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "plugin_excalidraw_v3_prototype"
+name = "plugin_excalidraw"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
-description = "Excalidraw prototype for the fused Lix Wasm Component API v3"
+description = "Excalidraw plugin for the fused Lix Wasm Component API v3"
 license.workspace = true
 homepage.workspace = true
 repository.workspace = true
@@ -21,6 +21,6 @@ crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 lix_order_key.workspace = true
-lix_plugin_api_v3_prototype = { workspace = true }
+lix_plugin_api = { workspace = true }
 serde_json = "1"
 uuid = { version = "1", features = ["std"] }
