+++ plugins/json/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "plugin_markdown_v3_prototype"
+name = "plugin_json"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
-description = "Markdown prototype for the fused Lix Wasm Component API v3"
+description = "JSON plugin for the fused Lix Wasm Component API v3"
 license.workspace = true
 homepage.workspace = true
 repository.workspace = true
@@ -21,11 +21,7 @@ crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 base64 = "0.22"
-chardetng = "1"
-encoding_rs = "0.8"
-lix_order_key.workspace = true
-lix_plugin_api_v3_prototype = { workspace = true }
-markdown-syntax = "=0.2.0"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
+blake3 = "1"
+lix_plugin_api = { workspace = true }
+serde_json = { version = "1", features = ["raw_value"] }
 uuid = { version = "1", features = ["std"] }
