+++ plugins/markdown/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "plugin_json_v3_prototype"
+name = "plugin_markdown"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
-description = "JSON prototype for the fused Lix Wasm Component API v3"
+description = "Markdown plugin for the fused Lix Wasm Component API v3"
 license.workspace = true
 homepage.workspace = true
 repository.workspace = true
@@ -21,7 +21,11 @@ crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 base64 = "0.22"
-blake3 = "1"
-lix_plugin_api_v3_prototype = { workspace = true }
-serde_json = { version = "1", features = ["raw_value"] }
+chardetng = "1"
+encoding_rs = "0.8"
+lix_order_key.workspace = true
+lix_plugin_api = { workspace = true }
+markdown-syntax = "=0.2.0"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
 uuid = { version = "1", features = ["std"] }
